extract aws_session
