+++ src/prowler/__init__.py
@@ -1,8 +1,8 @@
 #  Copyright (c) 2020 nalansitan.
 #  All rights reserved.
 
-import sys, getopt
-import boto3
+from prowler.settings import aws_session
+from prowler.checks import MFACheck
 
 name = 'prowler'
 __author__ = 'nalansitan'
@@ -10,32 +10,17 @@ __version__ = '0.0.1'
 
 
 def main():
-    args = get_argv(sys.argv[1:])
-    boto3.setup_default_session(profile_name=args['profile'])
     print("python version of prowler, continually updated...")
-    sts = boto3.client('sts')
+    sts = aws_session.client('sts')
     response = sts.get_caller_identity()
     print(response)
+    mfa_check = MFACheck().rules()
+    print(mfa_check)
+    # for check in mfa_check:
+    #     print(check.check_function())
 
 
-def print_help():
-    print('prowler -p <AWS_PROFILE>')
-
-
-def get_argv(argv):
-    result = {
-        'profile': 'default',
-    }
-    try:
-        opts, args = getopt.getopt(argv, "hp:", ["profile="])
-    except getopt.GetoptError:
-        print_help()
-        sys.exit(2)
-    for opt, arg in opts:
-        if opt == '-h':
-            print_help()
-            sys.exit()
-        elif opt in ("-p", "--profile"):
-            result['profile'] = opt
-    return result
+# debug
+if __name__ == '__main__':
+    main()
