+++ src/prowler/checks/AbstractCheck.py
