+++ src/prowler/checks/__init__.py
@@ -1,1 +1,2 @@
 from .AbstractCheck import AbstractCheck, Rule
+from .MFACheck import MFACheck
